修复正常场景下逻辑删除注解会出现的严重Bug
